+++ application/controllers/bid_details_controller.py
